refactor(vsock-host): split command tests
